refactored com.tutorials.hellotabwidget to com.topherlee.psonememorymanager
