+++ AndroidManifest.xml
@@ -8,7 +8,7 @@
 
 	
     <application android:icon="@drawable/memcard" android:label="@string/app_name">    	
-        <activity android:name="com.tutorials.hellotabwidget.MCTabsWidget"
+        <activity android:name="com.topherlee.psonememorymanager.MCTabsWidget"
                   android:label="@string/app_name"                  
                   android:screenOrientation="landscape">
             <intent-filter>
@@ -16,22 +16,22 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.tutorials.hellotabwidget.Card1Activity"  android:label="@string/app_name"
+        <activity android:name="com.topherlee.psonememorymanager.Card1Activity"  android:label="@string/app_name"
             android:screenOrientation="landscape"/> 
-    	<activity android:name="com.tutorials.hellotabwidget.Card2Activity"  android:label="@string/app_name"
+    	<activity android:name="com.topherlee.psonememorymanager.Card2Activity"  android:label="@string/app_name"
     	    android:screenOrientation="landscape"/>  
-    	<activity android:name="com.tutorials.hellotabwidget.FileBrowser" android:label="@string/app_name"
+    	<activity android:name="com.topherlee.psonememorymanager.FileBrowser" android:label="@string/app_name"
     	    android:screenOrientation="landscape"/>
-    	<activity android:name="com.tutorials.hellotabwidget.MCViewActivity"
+    	<activity android:name="com.topherlee.psonememorymanager.MCViewActivity"
     	    android:screenOrientation="landscape">
     	    <intent-filter>
             <action android:name="com.tutorials.hellotabwidget.COPYMC"></action>
         	</intent-filter>
     	</activity>
     	    
-    	<activity android:name="com.tutorials.hellotabwidget.MCViewActivityTab2"
+    	<activity android:name="com.topherlee.psonememorymanager.MCViewActivityTab2"
     	    android:screenOrientation="landscape"/>
-    	  	<activity android:name="com.tutorials.hellotabwidget.Prefs" android:screenOrientation="landscape"/>
-    	  	<activity android:name="com.tutorials.hellotabwidget.EditHeader" android:screenOrientation="landscape"/>
+    	  	<activity android:name="com.topherlee.psonememorymanager.Prefs" android:screenOrientation="landscape"/>
+    	  	<activity android:name="com.topherlee.psonememorymanager.EditHeader" android:screenOrientation="landscape"/>
     </application>
 </manifest>
