+++ src/com/topherlee/psonememorymanager/DescIconView.java
@@ -1,4 +1,4 @@
-package com.tutorials.hellotabwidget;
+package com.topherlee.psonememorymanager;
 
 import android.content.Context;
 import android.content.res.ColorStateList;
