+++ src/com/topherlee/psonememorymanager/MemoryCard.java
@@ -1,4 +1,4 @@
-package com.tutorials.hellotabwidget;
+package com.topherlee.psonememorymanager;
 
 import java.io.BufferedOutputStream;
 import java.io.DataInputStream;
