+++ src/com/topherlee/psonememorymanager/Prefs.java
@@ -1,4 +1,4 @@
-package com.tutorials.hellotabwidget;
+package com.topherlee.psonememorymanager;
 
 import java.util.Map;
